project/requirements.txt
django
psycopg2
django-location-field
six